+++ agents/src/voice/room_io/_output.ts
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 import type {
   AudioFrame,
-  LocalParticipant,
   Participant,
   RemoteParticipant,
   RemoteTrackPublication,
@@ -26,26 +25,7 @@ import {
 } from '../../constants.js';
 import { log } from '../../log.js';
 import { Future, Task } from '../../utils.js';
-
-function findMicrophoneTrackId(room: Room, identity: string): string {
-  let p: Participant | LocalParticipant | null = room.remoteParticipants.get(identity) ?? null;
-  if (identity === room.localParticipant?.identity) {
-    p = room.localParticipant;
-  }
-
-  if (p === null) {
-    throw new Error(`Participant ${identity} not found`);
-  }
-
-  for (const track of p.trackPublications.values()) {
-    if (track.source === TrackSource.SOURCE_MICROPHONE && track.sid) {
-      // find the first microphone track
-      return track.sid;
-    }
-  }
-
-  throw new Error(`Participant ${identity} does not have a microphone track`);
-}
+import { findMicrophoneTrackId } from '../transcription/index.js';
 
 export abstract class TextOutput {
   abstract captureText(text: string): Promise<void>;
